requirements.txt
streamlit
python-dotenv
google-generativeai
pymongo
bcrypt
python-jose
passlib
dnspython
certifi